Refactor Artist and ArtistList components
