+++ src/components/Artists/Artist.md
@@ -1,0 +1,15 @@
+One Artist:
+
+```js
+<Artist
+  artist="Unprocessed"
+  songTitle="Real"
+  id="a"
+  image="https://i.scdn.co/image/ab67616d0000b273251b8f4c6183988f5060b6eb"
+  toggleCurrentSongId={() => {}}
+  isSongPlaying={false}
+  currentSongId="a"
+  saveSong={() => {}}
+  savedSongs={[]}
+/>
+```
