--- src/components/Artists/ArtistList.md
@@ -1,19 +1,0 @@
-One list item:
-
-```js
-<ArtistList
-  artists={[
-    {
-      artist: 'Unprocessed',
-      songTitle: 'Real',
-      songSnippet: 'https://test/',
-      id: 'a',
-    },
-  ]}
-  toggleCurrentSongId={() => {}}
-  isSongPlaying={false}
-  currentSongId="a"
-  saveSong={() => {}}
-  SavedSongs={[]}
-/>
-```
